Update AssumeRole test to catch error that is now thrown in 4.0 for describe images when role has no permissions. Previous behavior was just to return nothing.
Update canned role test too. I used to derive the service endpoint url based on the ec2 endpoint. That used to end in /service/Eucalyptus now it ends in /service/compute. Since "compute" is shorter than "Eucalyptus" the derived url was not correct.

This change also moves uncompressing zerofile.gz from the compile ant target to the download_deps target. This is a convience for running vi an IDE such as IntelliJ rather than from commandline.